+++ Medidata.Cloud.ExcelLoader/ICellTypeValueConverterManager.cs
@@ -2,7 +2,7 @@ using DocumentFormat.OpenXml.Spreadsheet;
 
 namespace Medidata.Cloud.ExcelLoader
 {
-    public interface ICellTypeValueConverterFactory
+    public interface ICellTypeValueConverterManager
     {
         void GetCellTypeAndValue(object value, out CellValues cellType, out string cellValue);
         object GetCSharpValue(Cell cell);
