+++ Medidata.Rave.Tsdv.Loader.Sample/unity.config
@@ -21,15 +21,14 @@
       </constructor>
     </register>
 
-    <register type="Medidata.Cloud.ExcelLoader.ICellTypeValueConverterFactory, Medidata.Cloud.ExcelLoader"
-              mapTo="Medidata.Cloud.ExcelLoader.CellTypeValueConverterFactory, Medidata.Cloud.ExcelLoader"
-              name="converterFactory">
+    <register type="Medidata.Cloud.ExcelLoader.ICellTypeValueConverterManager, Medidata.Cloud.ExcelLoader"
+              mapTo="Medidata.Cloud.ExcelLoader.CellTypeValueConverterManager, Medidata.Cloud.ExcelLoader">
       <lifetime type="singleton" />
     </register>
     <register type="Medidata.Cloud.ExcelLoader.ISheetBuilder, Medidata.Cloud.ExcelLoader"
               mapTo="Medidata.Cloud.ExcelLoader.SheetBuilder, Medidata.Cloud.ExcelLoader">
       <constructor>
-        <param name="converterFactory" dependencyName="converterFactory" />
+        <param name="converterManager" dependencyType="Medidata.Cloud.ExcelLoader.ICellTypeValueConverterManager, Medidata.Cloud.ExcelLoader" />
         <param name="decorators">
           <array>
             <dependency type="Medidata.Cloud.ExcelLoader.SheetDecorators.HeaderSheetDecorator, Medidata.Cloud.ExcelLoader" />
@@ -44,11 +43,7 @@
       </constructor>
     </register>
     <register type="Medidata.Cloud.ExcelLoader.ISheetParser, Medidata.Cloud.ExcelLoader"
-              mapTo="Medidata.Cloud.ExcelLoader.SheetParser, Medidata.Cloud.ExcelLoader">
-      <constructor>
-        <param name="converterFactory" dependencyName="converterFactory" />
-      </constructor>
-    </register>
+              mapTo="Medidata.Cloud.ExcelLoader.SheetParser, Medidata.Cloud.ExcelLoader" />
     <register type="Medidata.Cloud.ExcelLoader.IExcelBuilder, Medidata.Cloud.ExcelLoader"
               mapTo="Medidata.Rave.Tsdv.Loader.CoveredExcelBuilder, Medidata.Rave.Tsdv.Loader" />
     <register type="Medidata.Cloud.ExcelLoader.IExcelParser, Medidata.Cloud.ExcelLoader"
